Fix submatch support (original: #17)
The original (ba01640936b2175cf2809f9ea0ab0443e4fb0472) would only match
the first invocation, and also was unsafe in patterns.

This patch breaks "end" patterns utilizing a backref to a Vim pattern
embedded within the referenced "start" submatch. I see this behavior as
pretty unexpected (and not that portable), so these sub-matches will now
be matched literally. This sanitation has a side-effect of resetting
interpretation to as if 'magic' was set, which is documented and feels
reasonable as a trade-off. (Detecting the last magic sequence isn't
free.) (E.g.: An "end" pattern '\vEnd #(\d+) of \1 \(line (\d+)\)' would
be written either '\vEnd #(\d+) of \1\v \(line (\d+)\)' (preferred) or
'\vEnd #(\d+) of \1 (line \(\d\+\))' (but this doesn't also work as a
normal Vim pattern).) This is documented, and relying on the hidden "\m"
is contraindicated (as this makes patterns interpret differently
everywhere else; the workaround of repeating yourself is benign).